+++ snow-spec/src/openapiNode.ts
